# Performance improvement for ProcessManager.Windows.IsProcessRunning