Performance optimization for the local machine:
First try to OpenProcess by id, if valid handle is returned, the process is definitely running.
Otherwise enumerate all processes and compare ids (the logic we had before)

Fix #17382

Benchmarks before:
```
              Method |      Mean |    StdDev |
-------------------- |---------- |---------- |
      GetProcessById | 3.5353 ms | 0.0693 ms |
   GetProcessesEqual | 3.1075 ms | 0.0720 ms |
GetProcessesNotEqual | 3.1015 ms | 0.0416 ms |
```
After:
```
              Method |      Mean |    StdDev |
-------------------- |---------- |---------- |
      GetProcessById | 1.4213 ms | 0.0581 ms |
   GetProcessesEqual | 3.1123 ms | 0.0615 ms |
GetProcessesNotEqual | 3.0089 ms | 0.0532 ms |
```

Test program:
```csharp
using System.Diagnostics;
using System.Linq;
using System.Reflection;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace GetProcessByIdBenchmark
{
    public class Program
    {
        private static int CurrentProcessId = Process.GetCurrentProcess().Id;
		
        [Benchmark]
        public void GetProcessById()
        {
            Process.GetProcessById(CurrentProcessId);
        }

        [Benchmark]
        public void GetProcessesEqual()
        {
            Process.GetProcesses().FirstOrDefault(a => a.Id == CurrentProcessId);
        }

        [Benchmark]
        public void GetProcessesNotEqual()
        {
            Process.GetProcesses().FirstOrDefault(a => a.Id != CurrentProcessId);
        }

        static void Main(string[] args)
        {
            BenchmarkSwitcher.FromAssembly(typeof(Program).GetTypeInfo().Assembly).Run(args);
        }
    }
}
```