@stephentoub 2017-03-24
Left a few comments, but otherwise LGTM.  Thank you for fixing this.

cc: @Priya91